自定义参数解析器

+++ src/main/kotlin/com/github/blanexie/vxph/torrent/dto/AnnounceReq.kt
@@ -3,7 +3,7 @@ package com.github.blanexie.vxph.torrent.dto
 
 class AnnounceReq(
     val peerId: String, val infoHash: String, val passKey: String, val left: Long, val downloaded: Long,
-    val uploaded: Long, val compact: Int, val event: String, val remoteAddr: String, val remotePort: Short
+    val uploaded: Long, val compact: Int, val event: String, val remoteAddr: String, val remotePort: Int
 ) {
